feat(navigator): ship the reference sandbox file-tool implementation
Moves the n2 file-tool engine out of the Cua cookbook into the SDK as
yutori.navigator.sandbox_tools: the in-VM FILE_TOOL_SCRIPT (python3 stdlib
only) plus ShellFileToolsMixin, which provides read/write/edit/grep/glob
over two adapter hooks (run_sandbox_shell, file_tool_cwd), and the shared
shell-result helpers (format_shell_output, render_image_result, truncation).
Contract strings are moved byte-exact — the cookbook's exact-string tests
pass unmodified with only import updates — so any sandbox adapter with a
shell and python3 can pin the current tool set without reimplementing the
tool contracts. The Cua adapter shrinks from 615 to ~240 lines.

Version 0.9.6 so single-file examples can depend on it from PyPI.
