+++ api.md
@@ -518,7 +518,7 @@ The keywords:
 | `max_steps` / `agent_timeout_seconds` | `None` | Turn and wall-clock budgets per `run()`/`resume()` call (`stopped_by == "max_steps"` / `"timeout"`). |
 | `compactor` | `"auto"` | `"auto"` attaches a fresh `N2InlineCompactor` (SDK 0.9.5+), so long runs are checkpointed instead of ending at the context limit. Pass `None` to disable, or an `N2Compactor` for a custom policy; it is called before each model request and the context-limit guard. Each applied compaction fires the `on_compaction` callback with `{"items_before", "items_after"}`. |
 
-Adapter hooks: a file handler (`read_file` and friends) may return `{"text": ..., "image_url": "data:..."}` so a `read` of an image file shows the model the image as well as the text; any computer handler that declares a `model_action=` keyword parameter receives the model's own call (`{"action": name, **arguments}`) alongside the translated arguments. Shell and file handlers own their result text end to end — see the reference implementations in `examples/navigator_n2/cua_adapter.py` and `examples/navigator_n2_daytona.py` for the formats n2 expects (`Exit code N` headers, `cat -n` line numbering, `[... output truncated, N more chars ...]` caps).
+Adapter hooks: a file handler (`read_file` and friends) may return `{"text": ..., "image_url": "data:..."}` so a `read` of an image file shows the model the image as well as the text; any computer handler that declares a `model_action=` keyword parameter receives the model's own call (`{"action": name, **arguments}`) alongside the translated arguments. Shell and file handlers own their result text end to end. The SDK ships the reference file-tool implementation: mix `yutori.navigator.ShellFileToolsMixin` into an adapter and implement two hooks — `run_sandbox_shell(command, *, timeout_seconds)` (any object with `stdout`/`stderr`/`returncode`) and `file_tool_cwd()` — and the sandbox-side `FILE_TOOL_SCRIPT` (python3 stdlib only) produces the exact formats n2 expects (`cat -n` line numbering, the sha256 read-before-edit gate, `[... output truncated, N more chars ...]` caps). `format_shell_output(output, exit_code)` renders `bash` results the same way (`Exit code N` headers). `render_image_result` turns an image `read` into visible image content. SDK 0.9.6+; `examples/navigator_n2/cua_adapter.py` shows the wiring.
 
 ### N2 context compaction
