+++ examples/navigator_n2/pyproject.toml
@@ -9,11 +9,11 @@ dependencies = [
     # create image-based cloud VMs. 0.1.17 keeps the local Docker runtime working;
     # uv then resolves cua-cli back to 0.1.13, the newest release without that floor.
     "cua-sandbox==0.1.17",
-    "yutori==0.9.5",
+    "yutori==0.9.6",
 ]
 
 [project.optional-dependencies]
-macos = ["yutori[macos]==0.9.5"]
+macos = ["yutori[macos]==0.9.6"]
 
 [tool.uv.sources]
 yutori = { path = "../..", editable = true }
