+++ llms.txt
@@ -165,7 +165,7 @@ The core action space covers clicks, scroll, type, key press, drag, mouse move/d
 
 ### Navigator n2 (computer use)
 
-Navigator n2 operates a full desktop. `N2ComputerAgent` defaults to `model="n2"`; pin the current tool set with `tool_set=TOOL_SET_COMPUTER_USE_LATEST` (`"computer_use_tools-20260825"`). It exposes `computer_batch`, `edit`, `read`, `write`, and `bash`. A batch holds up to 20 actions drawn from 15 GUI action types, runs sequentially against one observed frame, stops at its first error, and receives one screenshot result. n2 is non-streaming and rejects caller-provided `tools`, `disable_tools`, `json_schema`, `response_format`, and non-auto `tool_choice` — the pinned tool set is all-or-nothing, so implement every tool in it (the model runs shell work through `bash` rather than a GUI terminal, and expects `read` on an image file to return the image). A caller system message is appended under a `# User Instructions` header after the server's own prompt, which already tells the model the current US Pacific date/time and to resolve paths under `$HOME`. Send the full conversation; the server keeps images only in the two newest image-bearing messages. Long runs compact automatically by default (SDK 0.9.5+, `compactor="auto"` -> `N2InlineCompactor`); pass `compactor=None` to instead stop at the 128k context limit with `stopped_by="context_limit"`.
+Navigator n2 operates a full desktop. `N2ComputerAgent` defaults to `model="n2"`; pin the current tool set with `tool_set=TOOL_SET_COMPUTER_USE_LATEST` (`"computer_use_tools-20260825"`). It exposes `computer_batch`, `edit`, `read`, `write`, and `bash`. A batch holds up to 20 actions drawn from 15 GUI action types, runs sequentially against one observed frame, stops at its first error, and receives one screenshot result. n2 is non-streaming and rejects caller-provided `tools`, `disable_tools`, `json_schema`, `response_format`, and non-auto `tool_choice` — the pinned tool set is all-or-nothing, so implement every tool in it (the model runs shell work through `bash` rather than a GUI terminal, and expects `read` on an image file to return the image); SDK 0.9.6+ ships the reference file-tool implementation as `yutori.navigator.ShellFileToolsMixin` for any sandbox with a shell and python3. A caller system message is appended under a `# User Instructions` header after the server's own prompt, which already tells the model the current US Pacific date/time and to resolve paths under `$HOME`. Send the full conversation; the server keeps images only in the two newest image-bearing messages. Long runs compact automatically by default (SDK 0.9.5+, `compactor="auto"` -> `N2InlineCompactor`); pass `compactor=None` to instead stop at the 128k context limit with `stopped_by="context_limit"`.
 
 - **Local Docker:** [`examples/navigator_n2/`](https://github.com/yutori-ai/yutori-sdk-python/tree/main/examples/navigator_n2) implements the full current tool set in a disposable local container.
 - **Third-party Daytona desktop:** [`examples/navigator_n2_daytona.py`](https://github.com/yutori-ai/yutori-sdk-python/blob/main/examples/navigator_n2_daytona.py) is a compact hosted example. The SDK runs the loop; a Yutori-maintained `DaytonaComputer` adapter and lifecycle wiring execute actions on Daytona. The script declares Python 3.10+, `yutori>=0.9.4`, and `daytona==0.207.0` as inline dependencies, so run it with `uv run examples/navigator_n2_daytona.py "<task>"` plus a `DAYTONA_API_KEY`. This compact adapter intentionally pins the immutable `20260812` batch-plus-bash surface. Walkthrough: <https://docs.yutori.com/reference/n2-daytona>.
